split helper methods for building collections into separate classes
